@@ -2,7 +2,7 @@ class Api::V2::TeamsController < Api::V2::ApiController
 
   def me
     @teams = api_user.teams.order(:created_at)
-    @invites = api_user.team_invites.open
+    @invites = api_user.open_invites
   end
 
 end
