@@ -1,8 +1,16 @@
 # frozen_string_literal: true
 
 class TeamInviteController < ApplicationController
-  def new
+  before_action :check_team_member_rights, only: [:index, :create, :delete]
+
+  def index
     @team_invite_submissions = TeamInviteForm.new
+    @show_team_invite = TeamInvite.where(team_id: current_team).page(params[:page]).per(10)
+
+    respond_to do |format|
+      format.html
+      format.js
+    end
   end
 
   def create
@@ -12,10 +20,20 @@ class TeamInviteController < ApplicationController
     if @team_invite_submissions.valid?
       TeamInviteAdder.create_from_email_list(@team_invite_submissions.emails, @team)
       flash[:success] = 'Team invites have been sent!'
-      redirect_to manage_invites_path(team: @team.slug)
     else
-      render 'teams/manage'
+      flash[:error] = 'Invalid email(s)!'
     end
+    redirect_to manage_invites_path(team: @team.slug)
+  end
+
+  def delete
+    invite = TeamInvite.find(params[:id])
+    if invite.destroy
+      flash[:success] = 'Succesfully deleted invitation!'
+    else
+      flash[:error] = 'Something went wrong, please try again'
+    end
+    redirect_to manage_invites_path(team: current_team.slug)
   end
 
   def accept
