@@ -5,10 +5,14 @@ class UserMailer < ApplicationMailer
     suppress(Exception) {welcome_email(user).deliver_later}
   end
 
+  def invite_email(email, team)
+    @team = team
+
+    mail(to: email, subject: "You've been invited to join the Kudos-o-Matic")
+  end
+
   def welcome_email(user)
     @user = user
-
-    attachments.inline['logo.png'] = logo_attachment
 
     mail(to: user.email, subject: 'Welcome to the ₭udo-o-Matic!')
   end
@@ -16,15 +20,11 @@ class UserMailer < ApplicationMailer
   def export_start_email(user)
     @user = user
 
-    attachments.inline['logo.png'] = logo_attachment
-
     mail(to: user.email, subject: 'Started data export!')
   end
 
   def export_done_email(user, export)
     @user = user
-
-    attachments.inline['logo.png'] = logo_attachment
 
     @export = export
     mail(to: user.email, subject: 'Data export finished!')
