@@ -1,5 +1,45 @@
 # frozen_string_literal: true
 
+# == Schema Information
+#
+# Table name: users
+#
+#  id                     :integer          not null, primary key
+#  name                   :string
+#  created_at             :datetime         not null
+#  updated_at             :datetime         not null
+#  email                  :string           default(""), not null
+#  encrypted_password     :string           default(""), not null
+#  reset_password_token   :string
+#  reset_password_sent_at :datetime
+#  remember_created_at    :datetime
+#  sign_in_count          :integer          default(0), not null
+#  current_sign_in_at     :datetime
+#  last_sign_in_at        :datetime
+#  current_sign_in_ip     :inet
+#  last_sign_in_ip        :inet
+#  confirmation_token     :string
+#  confirmed_at           :datetime
+#  confirmation_sent_at   :datetime
+#  unconfirmed_email      :string
+#  failed_attempts        :integer          default(0), not null
+#  unlock_token           :string
+#  locked_at              :datetime
+#  provider               :string
+#  uid                    :string
+#  avatar_url             :string
+#  slack_name             :string
+#  admin                  :boolean          default(FALSE)
+#  api_token              :string
+#  deactivated_at         :datetime
+#  preferences            :json
+#  slack_id               :string
+#  slack_username         :string
+#  restricted             :boolean          default(FALSE)
+#  company_user           :boolean          default(FALSE)
+#
+
+
 class User < ActiveRecord::Base
   after_update :ensure_an_admin_remains
 
@@ -15,7 +55,6 @@ class User < ActiveRecord::Base
   has_many :received_transactions, class_name: 'Transaction', foreign_key: :receiver_id
   has_many :memberships, class_name: 'TeamMember', foreign_key: :user_id
   has_many :teams, through: :memberships
-  has_many :team_invites
   has_many :votes, foreign_key: 'voter_id'
   has_many :exports, foreign_key: 'user_id'
   has_many :fcm_tokens
@@ -94,8 +133,8 @@ class User < ActiveRecord::Base
     # TODO: when we implement the functionality to give/revoke admin rights, we need to make sure to invalidate @admin_rights[team.id]
   end
 
-  def invited_to?(team)
-    team_invites.open.where(team_id: team.id).any?
+  def open_invites
+    TeamInvite.where(email: self.email, accepted_at: nil, declined_at: nil)
   end
 
   def all_transactions
@@ -128,10 +167,6 @@ class User < ActiveRecord::Base
 
   def multiple_teams?
     teams.length > 1
-  end
-
-  def invites?
-    team_invites.length > 1
   end
 
   def slack_id_for_team(team)
