@@ -110,20 +110,6 @@ class SlackService
   def set_general_channel_id(team)
     return unless SLACK_IS_CONFIGURED
 
-    uri = URI.parse('https://slack.com/api/channels.list')
-    http = Net::HTTP.new(uri.host, uri.port)
-    http.use_ssl = true
-    request = Net::HTTP::Get.new(uri.path, 'Content-type' => 'application/json',
-                                           'Authorization' => "Bearer #{team.slack_bot_access_token}")
-
-    response = http.request(request).body
-
-    channels = JSON.parse(response)['channels']
-    channels.each do |c|
-      if c['name'] == 'general'
-        team.update_attribute(:channel_id, c['id'])
-        break
-      end
-    end
+    team.update_attribute(:channel_id, ENV['SLACK_CHANNEL'])
   end
 end
