@@ -4,7 +4,7 @@ require_relative 'application'
 # Initialize the Rails application.
 Rails.application.initialize!
 
-ActionMailer::Base.default :from => 'support@yourdomain.com'
+ActionMailer::Base.default :from => ENV['MAIL_USERNAME']
 
 # Remove field_with_error divs
 ActionView::Base.field_error_proc = Proc.new do |html_tag, instance|
