@@ -2,8 +2,9 @@ require 'rails_helper'
 require 'shared/controllers/admin/shared_expectations'
 
 RSpec.describe Admin::BalancesController, type: :controller do
-  let!(:team) { create :team }
+  let!(:team) {create :team}
   let!(:balance) {Balance.current(team)}
+  let!(:balance2) {Balance.create team: team, name: 'Balance 2'}
   let!(:user) {create(:user, :admin)}
 
   before do
@@ -11,30 +12,25 @@ RSpec.describe Admin::BalancesController, type: :controller do
   end
 
   describe 'PATCH #update' do
+
     let (:perform_update) {patch :update, params: {id: balance.id, balance: balance.attributes}}
 
-    context 'updating the a current balance to a normal balance' do
+    context 'updating the a current balance to a non-current balance' do
       before do
-        balance.current = false
+        # balance.current = false
       end
 
       context 'with at least one other current balance left' do
-        let!(:current_balance) {create(:balance, :current)}
         let! (:record_count_before_request) {Balance.count}
-
         before do
           perform_update
         end
 
-        it 'updates the current balance to a normal balance' do
-          expect(Balance.find(balance.id).current).to eq(false)
+        it 'cant remove the only current balance attached to this team' do
+          expect(Balance.find(balance.id).current).to eq(true)
         end
 
         expect_balance_count_same
-
-        it 'redirects back to the admin balance path' do
-          expect(response).to redirect_to(admin_balance_path)
-        end
       end
 
       context 'with no other current balances left' do
@@ -48,17 +44,16 @@ RSpec.describe Admin::BalancesController, type: :controller do
           expect(Balance.find(balance.id).current).to eq(true)
         end
 
-        expect_balance_count_same
-
         it 'redirects back to the edit admin balance path ' do
-          expect(response).to redirect_to(edit_admin_balance_path(balance.id))
+          expect(response).to redirect_to(admin_balance_path)
         end
       end
     end
   end
 
   describe 'DELETE #destroy' do
-    let (:perform_delete) {delete :destroy, params: {id: balance.id}}
+    let (:perform_delete) {delete :destroy, params: {id: balance2.id}}
+    let (:perform_delete_active) {delete :destroy, params: {id: balance.id}}
 
     context 'deleting the last current balance' do
       context 'with at least one other current balance left' do
@@ -70,7 +65,7 @@ RSpec.describe Admin::BalancesController, type: :controller do
         end
 
         it 'deletes the current balance' do
-          expect {Balance.find(balance.id)}.to raise_error(ActiveRecord::RecordNotFound)
+          expect {Balance.find(balance2.id)}.to raise_error(ActiveRecord::RecordNotFound)
         end
 
         expect_balance_count_decrease
@@ -84,7 +79,7 @@ RSpec.describe Admin::BalancesController, type: :controller do
         let! (:record_count_before_request) {Balance.count}
 
         before do
-          perform_delete
+          perform_delete_active
         end
 
         it 'does not delete the current balance' do
