@@ -14,34 +14,60 @@ RSpec.describe TeamMemberController, type: :controller do
   end
 
   describe 'GET #index' do
-    before do
-      sign_in user
-      get :index, team: team.slug
+    context 'as team manager' do
+      before do
+        sign_in user
+        get :index, team: team.slug
+      end
+
+      it 'gets index' do
+        expect(response).to render_template(:index)
+      end
+
+      it 'returns a 200 (ok) status code' do
+        expect(response).to have_http_status(200)
+      end
     end
 
-    it 'gets index' do
-      expect(response).to render_template(:index)
-    end
+    context 'as not team manager' do
+      before do
+        sign_in user2
+        get :index, team: team.slug
+      end
 
-    it 'returns a 200 (ok) status code' do
-      expect(response).to have_http_status(200)
+      it 'redirects to dashboard' do
+        expect(response).to redirect_to(dashboard_path(team: team.slug))
+      end
     end
   end
 
   describe 'DELETE #delete' do
     let!(:membership) { team.memberships.last }
 
-    before do
-      sign_in user
-      delete 'delete', id: membership.id, team: team.slug
+    context 'as team admin' do
+      before do
+        sign_in user
+        delete 'delete', id: membership.id, team: team.slug
+      end
+
+      it 'removes the membership' do
+        expect(membership.id).to_not eql(team.memberships.last.id)
+      end
+
+      it 'redirects to team members page' do
+        expect(response).to redirect_to(manage_team_members_path(team: team.slug))
+      end
     end
 
-    it 'removes the membership' do
-      expect(membership.id).to_not eql(team.memberships.last.id)
-    end
+    context 'as not team admin' do
+      before do
+        sign_in user2
+        delete 'delete', id: membership.id, team: team.slug
+      end
 
-    it 'redirects to team members page' do
-      expect(response).to redirect_to(manage_team_members_path(team: team.slug))
+      it 'redirects to dashboard' do
+        expect(response).to redirect_to(dashboard_path(team: team.slug))
+      end
     end
   end
 end
